# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "axum-ts-example",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/posts": {
      "get": {
        "tags": [
          "posts"
        ],
        "operationId": "list_posts",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status discriminant, e.g. `published`.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PostSort"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_PostResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "posts"
        ],
        "operationId": "create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/posts/{id}": {
      "get": {
        "tags": [
          "posts"
        ],
        "operationId": "get_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Post id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "posts"
        ],
        "operationId": "delete_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Post id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "posts"
        ],
        "operationId": "update_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Post id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/posts/{id}/comments": {
      "get": {
        "tags": [
          "comments"
        ],
        "operationId": "list_comments",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Post id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_Comment"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "comments"
        ],
        "operationId": "create_comment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Post id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "list_users",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_UserResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "create_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "get_user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Author": {
        "type": "object",
        "description": "Reused author summary embedded in posts and comments.",
        "required": [
          "id",
          "display_name"
        ],
        "properties": {
          "display_name": {
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        }
      },
      "Comment": {
        "type": "object",
        "required": [
          "id",
          "author",
          "text",
          "created_at"
        ],
        "properties": {
          "author": {
            "$ref": "#/components/schemas/Author"
          },
          "created_at": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "string"
          },
          "text": {
            "type": "string"
          }
        }
      },
      "CreateCommentRequest": {
        "type": "object",
        "required": [
          "text"
        ],
        "properties": {
          "text": {
            "type": "string"
          }
        }
      },
      "CreatePostRequest": {
        "type": "object",
        "required": [
          "title",
          "body"
        ],
        "properties": {
          "body": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "description": "Arbitrary metadata → `Record<string, string>` / `{ [key: string]: string }`.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "tags": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "Optional list → `string[] | undefined`."
          },
          "title": {
            "type": "string"
          }
        }
      },
      "CreateUserRequest": {
        "type": "object",
        "required": [
          "email",
          "display_name"
        ],
        "properties": {
          "display_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ListResponse_Comment": {
        "type": "object",
        "description": "Generic paginated wrapper. utoipa monomorphizes per instantiation\n(`ListResponse_PostResponse`, `ListResponse_UserResponse`, …).",
        "required": [
          "items"
        ],
        "properties": {
          "bookmark": {
            "type": [
              "string",
              "null"
            ]
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "author",
                "text",
                "created_at"
              ],
              "properties": {
                "author": {
                  "$ref": "#/components/schemas/Author"
                },
                "created_at": {
                  "type": "integer",
                  "format": "int64"
                },
                "id": {
                  "type": "string"
                },
                "text": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "ListResponse_PostResponse": {
        "type": "object",
        "description": "Generic paginated wrapper. utoipa monomorphizes per instantiation\n(`ListResponse_PostResponse`, `ListResponse_UserResponse`, …).",
        "required": [
          "items"
        ],
        "properties": {
          "bookmark": {
            "type": [
              "string",
              "null"
            ]
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "title",
                "body",
                "status",
                "author",
                "tags",
                "comments",
                "created_at"
              ],
              "properties": {
                "author": {
                  "$ref": "#/components/schemas/Author",
                  "description": "Nested object."
                },
                "body": {
                  "type": "string"
                },
                "comments": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Comment"
                  },
                  "description": "Nested array of objects."
                },
                "created_at": {
                  "type": "integer",
                  "format": "int64"
                },
                "id": {
                  "type": "string"
                },
                "status": {
                  "$ref": "#/components/schemas/PostStatus"
                },
                "tags": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "title": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "ListResponse_UserResponse": {
        "type": "object",
        "description": "Generic paginated wrapper. utoipa monomorphizes per instantiation\n(`ListResponse_PostResponse`, `ListResponse_UserResponse`, …).",
        "required": [
          "items"
        ],
        "properties": {
          "bookmark": {
            "type": [
              "string",
              "null"
            ]
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "email",
                "display_name",
                "role"
              ],
              "properties": {
                "display_name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "id": {
                  "type": "string"
                },
                "role": {
                  "$ref": "#/components/schemas/UserRole"
                }
              }
            }
          }
        }
      },
      "PostResponse": {
        "type": "object",
        "required": [
          "id",
          "title",
          "body",
          "status",
          "author",
          "tags",
          "comments",
          "created_at"
        ],
        "properties": {
          "author": {
            "$ref": "#/components/schemas/Author",
            "description": "Nested object."
          },
          "body": {
            "type": "string"
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            },
            "description": "Nested array of objects."
          },
          "created_at": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/PostStatus"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "title": {
            "type": "string"
          }
        }
      },
      "PostSort": {
        "type": "string",
        "description": "Plain enum used both as a field and as a query parameter → string union.",
        "enum": [
          "newest",
          "oldest",
          "popular"
        ]
      },
      "PostStatus": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "type"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "draft"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "published_at",
              "type"
            ],
            "properties": {
              "published_at": {
                "type": "integer",
                "format": "int64"
              },
              "type": {
                "type": "string",
                "enum": [
                  "published"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "reason",
              "type"
            ],
            "properties": {
              "reason": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "archived"
                ]
              }
            }
          }
        ],
        "description": "Tagged enum → discriminated union. Three variants incl. data-carrying ones."
      },
      "UpdatePostRequest": {
        "type": "object",
        "description": "PATCH semantics: every field optional → partial update payload.",
        "properties": {
          "body": {
            "type": [
              "string",
              "null"
            ]
          },
          "tags": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "required": [
          "id",
          "email",
          "display_name",
          "role"
        ],
        "properties": {
          "display_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        }
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "admin",
          "member",
          "guest"
        ]
      }
    }
  }
}